Initial commit, PHEWW...
This is like the 5th time I made a repository of the same project,
but hell I hope this will be the final one. Now a lot of things are
still not set here but what I'm trying to implement are the stuff that
I couldn't achieve in my previous attempts.

First, I couldn't get to have perfect collision detection in one of my
first few attempts.
Second, one of my other attempts involved putting the entire game logic
inside the matrix, which was pretty effective in simplifying some stuff
but as a result, this made it almost impossible to animate stuff.

I'm currently in the process to solve these 2 problems.

One final thing before I can finally proceed to actually settle in this
project is to implement input buffering which I still have no idea about
it but we'll see how this goes.