All the ghosts should be moving in random directions already
I know that the code is looking very2 messy right now, I'll just
refactor this later.

By the way, these ghost AIs only change directiosn when they hit a wall,
right?
I'm planning to define intersection tiles, so that instead of the ghosts
only changing directions when hitting a wall, I can instead define
intersection tiles so that they change directions randomly when they are
in a intersection tiles.

Another solution instead of intersection tiles would be for them to
continuously look for 90 degree direction change, it is kinda hard to
explain though