I finally made the AI a bit more smarter!
Look, yes I know that the code is messy asf, I'll work on that later but
I figured out how make the AI change directions when they are on an
intersection! I just made certain variables like "blinky_previous_pos"
to compare it to the current position of blinky, if they are not equal,
then blinky can change directions. It is pretty complicated though and
it is getting late rn so I don't really have that much time writing this
shit lol. This is getting a bit more interesting (again). I just hope
things will go a bit more smoother.