I just learned about raw pointers while learning some C and I used it when pacman "eats" the food and power pellets.
Not sure if this is a faster way to do stuff, probably not that much, I
just changed the food_eat method to accept a pointer instead and then do
in-place mutation of the game_map (in a more unsafe manner though). But
still, it seems fun to experiment with it because I just recently learned
about pointers and memory management.